+++ TestEnvironment.psm1
@@ -164,7 +164,13 @@ Export-ModuleMember -Function @(
     'New-FreeIPASudoRule',
     'New-FreeIPARole',
     'New-FreeIPAPasswordPolicy',
-    'New-FreeIPAService'
+    'New-FreeIPAService',
+    'New-FreeIPAIdView',
+    'New-FreeIPAOtpToken',
+    'New-FreeIPAAutomemberRule',
+    'New-FreeIPAAutomount',
+    'New-FreeIPASelinuxUserMap',
+    'New-FreeIPACertMapRule'
 )
 
 $ExecutionContext.SessionState.Module.OnRemove = {